requirements.txt
Django==2.2.19
pytz==2022.1
sqlparse==0.4.2
python-dotenv==0.20.0
pillow
sorl-thumbnail